Move MessageService, return IUserMessage from the type reader
