+++ Modix.Bot/MessageTypeReader.cs
@@ -2,7 +2,7 @@
 using System.Threading.Tasks;
 using Discord.Commands;
 using Microsoft.Extensions.DependencyInjection;
-using Modix.Services.Messages;
+using Modix.Services.Core;
 
 namespace Modix
 {
@@ -12,9 +12,9 @@ namespace Modix
         {
             if (ulong.TryParse(input, out var messageId))
             {
-                var giveawayService = services.GetRequiredService<IMessageService>();
+                var messageService = services.GetRequiredService<IMessageService>();
 
-                var message = await giveawayService.FindMessageInUnknownChannelAsync(context.Guild.Id, messageId);
+                var message = await messageService.FindMessageAsync(context.Guild.Id, messageId);
 
                 if (message is { })
                     return TypeReaderResult.FromSuccess(message);
