+++ Modix.Services/Core/CoreSetup.cs
@@ -49,6 +49,7 @@ namespace Modix.Services.Core
                 .AddScoped<IDesignatedChannelMappingRepository, DesignatedChannelMappingRepository>()
                 .AddScoped<IDesignatedRoleMappingRepository, DesignatedRoleMappingRepository>()
                 .AddScoped<IMessageRepository, MessageRepository>()
-                .AddScoped<IGuildService, GuildService>();
+                .AddScoped<IGuildService, GuildService>()
+                .AddScoped<IMessageService, MessageService>();
     }
 }
